Refactor common webpack config
